src/requirements.txt
streamlit
playwright
playwright-stealth
beautifulsoup4
langchain-openai
python-dotenv